// packages/config/src/schema.d.ts
import { z } from 'zod';
export declare const envSchema: z.ZodObject<{
    NODE_ENV: z.ZodDefault<z.ZodEnum<["development", "test", "staging", "production"]>>;
    APP_NAME: z.ZodDefault<z.ZodString>;
    PORT: z.ZodDefault<z.ZodNumber>;
    WEB_PORT: z.ZodDefault<z.ZodNumber>;
    API_BASE_URL: z.ZodDefault<z.ZodString>;
    WEB_BASE_URL: z.ZodDefault<z.ZodString>;
    DATABASE_URL: z.ZodDefault<z.ZodString>;
    DATABASE_POOL_MIN: z.ZodDefault<z.ZodNumber>;
    DATABASE_POOL_MAX: z.ZodDefault<z.ZodNumber>;
    REDIS_HOST: z.ZodDefault<z.ZodString>;
    REDIS_PORT: z.ZodDefault<z.ZodNumber>;
    REDIS_PASSWORD: z.ZodDefault<z.ZodOptional<z.ZodString>>;
    REDIS_DB: z.ZodDefault<z.ZodNumber>;
    SESSION_SECRET: z.ZodDefault<z.ZodString>;
    SESSION_TTL_HOURS: z.ZodDefault<z.ZodNumber>;
    SESSION_COOKIE_NAME: z.ZodDefault<z.ZodString>;
    COOKIE_DOMAIN: z.ZodDefault<z.ZodString>;
    COOKIE_SECURE: z.ZodDefault<z.ZodEffects<z.ZodString, boolean, string>>;
    CSRF_SECRET: z.ZodDefault<z.ZodString>;
    MFA_APP_NAME: z.ZodDefault<z.ZodString>;
    MFA_BACKUP_CODES_COUNT: z.ZodDefault<z.ZodNumber>;
    LOG_LEVEL: z.ZodDefault<z.ZodEnum<["fatal", "error", "warn", "info", "debug", "trace"]>>;
    LOG_PRETTY: z.ZodDefault<z.ZodEffects<z.ZodString, boolean, string>>;
    CORS_ALLOWED_ORIGINS: z.ZodEffects<z.ZodDefault<z.ZodString>, string[], string | undefined>;
}, "strip", z.ZodTypeAny, {
    LOG_PRETTY: boolean;
    NODE_ENV: "development" | "test" | "staging" | "production";
    LOG_LEVEL: "fatal" | "error" | "warn" | "info" | "debug" | "trace";
    APP_NAME: string;
    PORT: number;
    WEB_PORT: number;
    API_BASE_URL: string;
    WEB_BASE_URL: string;
    DATABASE_URL: string;
    DATABASE_POOL_MIN: number;
    DATABASE_POOL_MAX: number;
    REDIS_HOST: string;
    REDIS_PORT: number;
    REDIS_PASSWORD: string;
    REDIS_DB: number;
    SESSION_SECRET: string;
    SESSION_TTL_HOURS: number;
    SESSION_COOKIE_NAME: string;
    COOKIE_DOMAIN: string;
    COOKIE_SECURE: boolean;
    CSRF_SECRET: string;
    MFA_APP_NAME: string;
    MFA_BACKUP_CODES_COUNT: number;
    CORS_ALLOWED_ORIGINS: string[];
}, {
    LOG_PRETTY?: string | undefined;
    NODE_ENV?: "development" | "test" | "staging" | "production" | undefined;
    LOG_LEVEL?: "fatal" | "error" | "warn" | "info" | "debug" | "trace" | undefined;
    APP_NAME?: string | undefined;
    PORT?: number | undefined;
    WEB_PORT?: number | undefined;
    API_BASE_URL?: string | undefined;
    WEB_BASE_URL?: string | undefined;
    DATABASE_URL?: string | undefined;
    DATABASE_POOL_MIN?: number | undefined;
    DATABASE_POOL_MAX?: number | undefined;
    REDIS_HOST?: string | undefined;
    REDIS_PORT?: number | undefined;
    REDIS_PASSWORD?: string | undefined;
    REDIS_DB?: number | undefined;
    SESSION_SECRET?: string | undefined;
    SESSION_TTL_HOURS?: number | undefined;
    SESSION_COOKIE_NAME?: string | undefined;
    COOKIE_DOMAIN?: string | undefined;
    COOKIE_SECURE?: string | undefined;
    CSRF_SECRET?: string | undefined;
    MFA_APP_NAME?: string | undefined;
    MFA_BACKUP_CODES_COUNT?: number | undefined;
    CORS_ALLOWED_ORIGINS?: string | undefined;
}>;
export type EnvConfig = z.infer<typeof envSchema>;
//# sourceMappingURL=schema.d.ts.map